% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_sensor_cols}
\alias{.get_sensor_cols}
\title{Internal: identify sensor columns}
\usage{
.get_sensor_cols(data, sensor_cols = NULL, exclude = "timestamp")
}
\description{
Internal: identify sensor columns
}
\keyword{internal}
