% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.pick_timestamp_col}
\alias{.pick_timestamp_col}
\title{Internal: pick timestamp column}
\usage{
.pick_timestamp_col(data, timestamp_col = "timestamp")
}
\description{
Internal: pick timestamp column
}
\keyword{internal}
