% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.simple_iso_score_1d}
\alias{.simple_iso_score_1d}
\title{Internal: simple isolation score for one numeric vector}
\usage{
.simple_iso_score_1d(x, ntrees = 40L)
}
\description{
Internal: simple isolation score for one numeric vector
}
\keyword{internal}
